+++ inlang/source-code/cli/src/commands/project/validate.ts
@@ -4,21 +4,15 @@ import { log } from "../../utilities/log.js"
 
 export const validate = new Command()
 	.command("validate")
-	.description("Validate the inlang config.")
+	.description("Validate the inlang prokect settings file.")
 	.action(validateCommandAction)
 
 export async function validateCommandAction() {
 	try {
-		log.info("🔎 Validating the config file...")
-
-		// Get the config
-		const { error } = await getInlangProject()
-		if (error) {
-			log.error(`${error} (${error.message})`)
-			return
-		}
-
-		log.success("Inlang config is valid!")
+		log.info("🔎 Validating the project settings file...")
+		// if `getInlangProject` doesn't return, the project is valid
+		await getInlangProject()
+		log.success("The project settings file is valid!")
 	} catch (error) {
 		log.error(error)
 	}