+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/example/src/routes/+layout.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  import ParaglideJsSvelteKitAdapter from "../lib/ParaglideSvelteKit.svelte"
+</script>
+
+<ParaglideJsSvelteKitAdapter>
+  <slot></slot>
+</ParaglideJsSvelteKitAdapter>