+++ inlang/source-code/website/src/components/Chip.tsx
@@ -6,9 +6,9 @@ export function Chip(props: { text: string | undefined; color: string; customCla
 		<div
 			style={{
 				color: props.color,
-				background: `${props.color}25`,
+				background: `${props.color}16`,
 			}}
-			class={`text-sm px-2 py-1 rounded-full ${props.customClasses}`}
+			class={`text-xs px-2 py-[2px] rounded-full ${props.customClasses}`}
 		>
 			{props.text}
 		</div>