+++ inlang/source-code/website/src/pages/newsletter/index.page.tsx
@@ -1,7 +1,7 @@
-import { Layout } from "../Layout.jsx"
 import { useI18n } from "@solid-primitives/i18n"
 import { Meta, Title } from "@solidjs/meta"
 import { NewsletterForm } from "#src/components/NewsletterForm.jsx"
+import MarketplaceLayout from "#src/components/marketplace/MarketplaceLayout.jsx"
 
 export type PageProps = {
 	markdown: string
@@ -19,7 +19,7 @@ export function Page() {
 			/>
 			<Meta name="og:image" content="/images/inlang-social-image.jpg" />
 			<Meta name="robots" content="noindex" />
-			<Layout>
+			<MarketplaceLayout>
 				<div class="relative max-w-screen-xl w-full mx-auto bg-background">
 					<div class="invisible xl:visible absolute top-0 left-0 h-full w-full z-0">
 						<div class="flex w-full h-full justify-between mx-auto items-end">
@@ -52,7 +52,7 @@ export function Page() {
 						</div>
 					</div>
 				</div>
-			</Layout>
+			</MarketplaceLayout>
 		</>
 	)
 }