+++ inlang/source-code/message-lint-rules/messageWithoutSource/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inlang/message-lint-rule-without-source
 
+## 1.4.0
+
+### Minor Changes
+
+- 0055f20b1: update README
+
 ## 1.3.0
 
 ### Minor Changes